Change RES_NOTSTORED to RES_NOTFOUND
If a key wasn't found when deleting, Memcached returns `RES_NOTFOUND`. I'm not sure that a `delete` operation would ever return `RES_NOTSTORED`.

In our case, this led to code like:
```php
$keys = ['11211:1', '35603:1'];
$memcached->delete($keys);
```

If the key `11211:1` doesn't exist, it will prevent `35603:1` from being deleted as well, which imo, violates the principal of least surprise.